添加到出账批号、在出账编号移出完成

+++ SocialSecurity/SocCommandService/SocCommandService-Entity/src/main/java/com/ciicsh/gto/afsupportcenter/socialsecurity/soccommandservice/entity/SsPayment.java
@@ -15,7 +15,7 @@ import java.io.Serializable;
  * </p>
  *
  * @author HuangXing
- * @since 2017-12-26
+ * @since 2017-12-27
  */
 @TableName("ss_payment")
 public class SsPayment implements Serializable {
@@ -23,7 +23,7 @@ public class SsPayment implements Serializable {
     private static final long serialVersionUID = 1L;
 
 	@TableId(value="payment_id", type= IdType.AUTO)
-	private Integer paymentId;
+	private Long paymentId;
     /**
      * 出账批号
      */
@@ -141,11 +141,11 @@ public class SsPayment implements Serializable {
 	private String modifiedBy;
 
 
-	public Integer getPaymentId() {
+	public Long getPaymentId() {
 		return paymentId;
 	}
 
-	public void setPaymentId(Integer paymentId) {
+	public void setPaymentId(Long paymentId) {
 		this.paymentId = paymentId;
 	}
